requirements.txt
flask==2.3.3
python-dotenv==1.0.0
groq==0.4.0
SpeechRecognition==3.10.0
Pillow==11.2.1
pydub==0.25.1
Werkzeug==2.3.7
